+++ rts/Sim/Path/IPathController.cpp
@@ -79,7 +79,8 @@ short GMTDefaultPathController::GetDeltaHeading(
 	// add lookahead term to avoid overshooting target heading
 	// note that turnBrakeDist is always positive
 	const short brakeDistFactor = (absTurnSpeed >= maxTurnAccel);
-	const short stopTurnHeading = oldHeading + (turnBrakeDist * Sign(curTurnSpeed) * brakeDistFactor);
+	// via int, a direct float->short is UB once out of range (#3075)
+	const short stopTurnHeading = short(int(oldHeading + (turnBrakeDist * Sign(curTurnSpeed) * brakeDistFactor)));
 	const short curDeltaHeading = newHeading - stopTurnHeading;
 
 	if (brakeDistFactor == 0) {
